+++ composite/apy-finance/src/adapter.ts
@@ -1,6 +1,6 @@
 import { Validator } from '@chainlink/external-adapter'
 import { AdapterResponse, Execute, AdapterRequest } from '@chainlink/types'
-import TokenAllocation from '@chainlink/token-allocation-adapter'
+import * as TokenAllocation from '@chainlink/token-allocation-adapter'
 import makeRegistry from './registry'
 import { makeConfig, Config } from './config'
 
@@ -8,14 +8,11 @@ export const execute = async (input: AdapterRequest, config: Config): Promise<Ad
   const validator = new Validator(input)
   if (validator.error) throw validator.error
 
-  const registry = await makeRegistry(config.addressRegistry, config.rpcUrl)
+  const registry = await makeRegistry(config.registryAddr, config.rpcUrl)
   const allocations = await registry.getAllocations()
 
-  const tokenAllocationExecute = TokenAllocation.makeExecute()
-
-  return await tokenAllocationExecute({
-    data: { ...input.data, ...allocations },
-  })
+  const _execute = TokenAllocation.makeExecute()
+  return await _execute({ id: validator.validated.id, data: { ...input.data, allocations } })
 }
 
 export const makeExecute = (config?: Config): Execute => {