+++ composite/apy-finance/src/registry/index.ts
@@ -1,39 +1,48 @@
-import { ethers } from 'ethers'
-import registryAbi from './IRegistry.json'
-import assetAllocationAbi from './IAssetAllocation.json'
+import { BigNumber, ethers } from 'ethers'
+import registryAbi from '../abi/IRegistry.json'
+import assetAllocationAbi from '../abi/IAssetAllocation.json'
+import erc20 from '@openzeppelin/contracts/build/contracts/ERC20.json'
+import { types } from '@chainlink/token-allocation-adapter'
 
-const getRegistry = (address: string, rpcUrl: string, abi: any): ethers.Contract => {
-  const provider = new ethers.providers.JsonRpcProvider(rpcUrl)
-  return new ethers.Contract(address, abi, provider)
-}
+export type GetAllocations = (
+  registry: ethers.Contract,
+  decimalsOf: (address: string) => Promise<number>,
+) => () => Promise<types.TokenAllocations>
 
-type Allocations = {
-  components: string[]
-  units: string[]
-}
-
-export type GetAllocations = (registry: ethers.Contract) => () => Promise<Allocations>
-
-const getAllocations: GetAllocations = (registry) => async () => {
+const getAllocations: GetAllocations = (registry, decimalsOf) => async () => {
   const tokenAddresses = await registry.getTokenAddresses()
-  const [components, units]: string[][] = await Promise.all([
+  const [components, balances, decimals]: any = await Promise.all([
     Promise.all(tokenAddresses.map((address: string) => registry.symbolOf(address))),
     Promise.all(tokenAddresses.map((address: string) => registry.balanceOf(address))),
+    Promise.all(tokenAddresses.map(async (address: string) => decimalsOf(address))),
   ])
-  return { components, units }
+
+  return components.map((symbol: string, i: number) => ({
+    symbol,
+    balance: BigNumber.from(balances[i]),
+    decimals: decimals[i],
+  }))
 }
 
 type Registry = {
-  getAllocations: () => Promise<Allocations>
+  getAllocations: () => Promise<types.TokenAllocations>
 }
 
 const makeRegistry = async (address: string, rpcUrl: string): Promise<Registry> => {
-  const registry = getRegistry(address, rpcUrl, registryAbi)
+  const provider = new ethers.providers.JsonRpcProvider(rpcUrl)
+  const registry = new ethers.Contract(address, registryAbi, provider)
   const chainlinkRegistryAddress = await registry.chainlinkRegistryAddress()
-  const chainlinkRegistry = getRegistry(chainlinkRegistryAddress, rpcUrl, assetAllocationAbi)
+  const chainlinkRegistry = new ethers.Contract(
+    chainlinkRegistryAddress,
+    assetAllocationAbi,
+    provider,
+  )
+
+  const _getERC20 = (address: string) => new ethers.Contract(address, erc20.abi, provider)
+  const _decimalsOf = (address: string) => _getERC20(address).decimals()
 
   return {
-    getAllocations: getAllocations(chainlinkRegistry),
+    getAllocations: getAllocations(chainlinkRegistry, _decimalsOf),
   }
 }