+++ dydx-stark/src/endpoint/starkex.ts
@@ -1,8 +1,9 @@
 import assert from 'assert'
-import BN from 'bn.js'
-import { ethers } from 'ethers'
+import { ethers, BigNumber } from 'ethers'
 import * as starkwareCrypto from '@authereum/starkware-crypto'
-import { AdapterError } from '@chainlink/external-adapter'
+import { logger, AdapterError } from '@chainlink/external-adapter'
+import { util } from '@chainlink/ea-bootstrap'
+import { Decimal } from 'decimal.js'
 
 export type PriceDataPoint = {
   oracleName: string
@@ -19,17 +20,15 @@ export type PriceStarkPayload = PriceDataPoint & {
 
 const MAX_DECIMALS = 18
 
-const ZERO_BN = new BN('0')
-const TWO_BN = new BN('2')
-const TEN_BN = new BN('10')
+const ZERO_BN = BigNumber.from('0')
+const TWO_BN = BigNumber.from('2')
 
-const powOfTwo = (num: number) => TWO_BN.pow(new BN(num))
-const powOfTen = (num: number) => TEN_BN.pow(new BN(num))
+const powOfTwo = (num: number) => TWO_BN.pow(num)
 
-const ERROR_MSG_PRICE_NEGATIVE = 'Price must be a positive number.'
-const ERROR_MSG_PRICE_PRECISION_LOSS =
+const ERROR_PRICE_NEGATIVE = 'Price must be a positive number.'
+const WARN_PRECISION_LOSS_NUMBER =
   'Please use string type to avoid precision loss with very small/big numbers.'
-const ERROR_MSG_PRICE_MAX_DECIMALS = 'Price has too many decimals.'
+const WARN_PRECISION_LOSS_STRING = `Precision loss detected: over ${MAX_DECIMALS} decimals.`
 
 const error400 = (message: string) => new AdapterError({ message, statusCode: 400 })
 
@@ -45,46 +44,25 @@ const error400 = (message: string) => new AdapterError({ message, statusCode: 40
 export const requireNormalizedPrice = (price: number | string): string => {
   // Check if negative number
   if (isNaN(Number(price)) || Number(price) < 0) {
-    throw error400(`${ERROR_MSG_PRICE_NEGATIVE} Got: ${price}`)
+    throw error400(`${ERROR_PRICE_NEGATIVE} Got: ${price}`)
   }
+
+  const _normalize = (num: number | string): Decimal => new Decimal(num).mul(10 ** MAX_DECIMALS)
+
+  const _tooMuchPrecision = (num: number | string): boolean =>
+    util.toFixedMax(_normalize(num), 1).indexOf('.') !== -1
 
   // Check if there is any loss of precision
   if (typeof price === 'number') {
     // TODO: more precision loss detection with floats
     const overSafeValue = price > Number.MAX_SAFE_INTEGER
-    if (overSafeValue) {
-      throw error400(`${ERROR_MSG_PRICE_PRECISION_LOSS} Got: ${price}.`)
-    }
+    if (overSafeValue || _tooMuchPrecision(price))
+      logger.warn(`${WARN_PRECISION_LOSS_NUMBER} Got: ${price}`)
+  } else if (_tooMuchPrecision(price)) {
+    logger.warn(`${WARN_PRECISION_LOSS_STRING} Got: ${price}`)
   }
 
-  // Convert number to decimal string (no scientific notation)
-  const _toString = (n: number) => {
-    const nStr = n.toString()
-    const isScientificNotation = nStr.indexOf('e') !== -1
-    if (!isScientificNotation) return nStr
-    return (
-      n
-        .toFixed(MAX_DECIMALS)
-        // remove trailing zeros
-        .replace(/(\.\d*?[1-9])0+$/g, '$1')
-        // remove decimal part if all zeros (or only decimal point)
-        .replace(/\.0*$/g, '')
-    )
-  }
-
-  const priceStr = typeof price === 'number' ? _toString(price as number) : (price as string)
-  const priceStrParts = priceStr.split('.')
-  const priceBig = new BN(priceStrParts[0]).mul(powOfTen(MAX_DECIMALS))
-  const decimals = (priceStrParts[1] && priceStrParts[1].length) || 0
-
-  if (decimals === 0) return priceBig.toString()
-  // Check if too many decimals
-  if (decimals > MAX_DECIMALS) {
-    throw error400(`${ERROR_MSG_PRICE_MAX_DECIMALS} Got: ${decimals}; Max: ${MAX_DECIMALS}`)
-  }
-
-  const decimalValBig = new BN(priceStrParts[1]).mul(powOfTen(MAX_DECIMALS - decimals))
-  return priceBig.add(decimalValBig).toString()
+  return _normalize(price).toFixed(0)
 }
 
 /**
@@ -103,10 +81,10 @@ export const getPricePayload = async (
   const keyPair = await getKeyPair(privateKey, starkMessage)
 
   // 4. Hash the required parameters
-  const message = getPriceMessage(data)
+  const message = getPriceMessage(data).toHexString().substr(2)
 
   // 5. Sign with your private stark key and the hash message to get r,s
-  const { r, s } = starkwareCrypto.sign(keyPair, message.toString(16))
+  const { r, s } = starkwareCrypto.sign(keyPair, message)
 
   // 6. Generate the public key (pub_key) with your private key
   const starkPublicKey = starkwareCrypto.getStarkPublicKey(keyPair)
@@ -138,7 +116,7 @@ export const getKeyPair = async (
   const hash = ethers.utils.keccak256(flatSig)
 
   // 3. Cut the last 5 bits of it to get your 251-bit-long private stark key
-  const pk = new BN(hash.substr(2), 16).shrn(5).toString(16)
+  const pk = BigNumber.from(hash).shr(5).toHexString().substr(2)
 
   return starkwareCrypto.getKeyPair(pk)
 }
@@ -148,17 +126,17 @@ export const getKeyPair = async (
  *
  * @param data price data point to hash
  */
-export const getPriceMessage = (data: PriceDataPoint): BN => {
+export const getPriceMessage = (data: PriceDataPoint): BigNumber => {
   // padded to 40 bit
-  const hexOracleName = Buffer.from(data.oracleName).toString('hex').padEnd(10, '0')
+  const hexOracleName = '0x' + Buffer.from(data.oracleName).toString('hex').padEnd(10, '0')
   // padded to 128 bit
-  const hexAssetName = Buffer.from(data.assetName).toString('hex').padEnd(32, '0')
+  const hexAssetName = '0x' + Buffer.from(data.assetName).toString('hex').padEnd(32, '0')
 
   return getPriceMessageRaw(
-    new BN(hexOracleName, 16),
-    new BN(hexAssetName, 16),
-    new BN(data.timestamp),
-    new BN(data.price),
+    BigNumber.from(hexOracleName),
+    BigNumber.from(hexAssetName),
+    BigNumber.from(data.timestamp),
+    BigNumber.from(data.price),
   )
 }
 
@@ -182,7 +160,12 @@ export const getPriceMessage = (data: PriceDataPoint): BN => {
  * @param timestamp a 32 bit number, represents seconds since epoch
  * @param price a 120-bit number
  */
-const getPriceMessageRaw = (oracleName: BN, assetName: BN, timestamp: BN, price: BN): BN => {
+const getPriceMessageRaw = (
+  oracleName: BigNumber,
+  assetName: BigNumber,
+  timestamp: BigNumber,
+  price: BigNumber,
+): BigNumber => {
   assert(oracleName.gte(ZERO_BN), error400('oracleName must be >= 0'))
   assert(oracleName.lt(powOfTwo(40)), error400('oracleName must be < 2 ** 40'))
 
@@ -196,14 +179,14 @@ const getPriceMessageRaw = (oracleName: BN, assetName: BN, timestamp: BN, price:
   assert(price.lt(powOfTwo(120)), error400('price must be < 2 ** 120'))
 
   // The first number to hash is the oracle name (Maker) and the asset name.
-  const first_number = assetName.shln(40).add(oracleName)
+  const first_number = assetName.shl(40).add(oracleName)
 
   // The second number is timestamp in the 32 LSB, then the price.
-  const second_number = price.shln(32).add(timestamp)
+  const second_number = price.shl(32).add(timestamp)
 
-  const w1 = first_number.toString(16)
-  const w2 = second_number.toString(16)
+  const w1 = first_number.toHexString().substr(2)
+  const w2 = second_number.toHexString().substr(2)
   const hash = starkwareCrypto.hashMessage([w1, w2])
 
-  return new BN(hash, 16)
+  return BigNumber.from('0x' + hash)
 }