+++ token-allocation/src/index.ts
@@ -1,7 +1,0 @@
-import { expose, util } from '@chainlink/ea-bootstrap'
-import { makeExecute } from './adapter'
-import { makeConfig } from './config'
-
-const NAME = 'Token-Allocations'
-
-export = { NAME, makeExecute, makeConfig, ...expose(util.wrapExecute(makeExecute())) }